djangoProject/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.1.0
psycopg2-binary==2.8.6
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
